+++ src/components/controls/RasterExpression.svelte
@@ -2,7 +2,7 @@
   import Fa from 'svelte-fa'
 
   import type { Layer } from '$lib/types'
-  import RasterRefineContainer from '$components/RasterRefineContainer.svelte'
+  import RasterRefineContainer from '$components/controls/RasterRefineContainer.svelte'
   import RasterExpressionSimple from '$components/controls/RasterExpressionSimple.svelte'
   import { faThumbsUp } from '@fortawesome/free-solid-svg-icons/faThumbsUp'
   import { faMagnifyingGlassPlus } from '@fortawesome/free-solid-svg-icons/faMagnifyingGlassPlus'
